New refactoring scripts. manufacturer_prefix works like a scrape except afte rthe fact. It strips the prefix form the name and moves the file. It doesn't leave a symlink so the spider should be updated in the same way. set_manufacturer moves a file, leaves a symlink and sets the manufacturer field. This is useful for moving unknown manufacturer parts to the appropriate directory if they are created by the site scraped.

+++ refactor_scripts/merge.py
@@ -21,8 +21,10 @@ def merge(original, new):
 def mergeFiles(merges, destination, part = None, test = False):
   if os.path.isfile(destination):
     with open(destination, "r") as f:
-      part = json.load(f)
+      part.update(json.load(f))
   for fn in merges:
+    if os.path.islink(fn):
+      continue
     with open(fn, "r") as f:
       this_part = json.load(f)
       if part:
@@ -32,7 +34,7 @@ def mergeFiles(merges, destination, part = None, test = False):
 
   for fn in merges:
     target = os.path.relpath(destination, os.path.dirname(fn))
-    if not test:
+    if not test and fn != target:
       os.remove(fn)
       os.symlink(target, fn)
     else:
